switched application properties to configuration for local usage, added null ip handling

+++ src/main/java/com/example/gradledemo/controllers/DeleteFileController.java
@@ -6,6 +6,7 @@ import com.example.gradledemo.persistence.SystemItemRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.lang.Nullable;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
@@ -18,9 +19,12 @@ public class DeleteFileController {
     @Autowired
     SystemItemService deleteService;
 
-    @DeleteMapping("/delete/{id}")
+    @DeleteMapping({"/delete/{id}", "/delete/"})
     @ResponseBody
-    public ResponseEntity<HashMap<String, Object>> deleteFiles(@PathVariable String id) {
+    public ResponseEntity<HashMap<String, Object>> deleteFiles(@PathVariable @Nullable String id) {
+        if(id == null) {
+            return deleteService.code400Response();
+        }
         return deleteService.deleteFilesResponse(id);
     }
 }
